Renames LZ4 to CodecLz4

+++ src/lz4frame.jl
@@ -94,7 +94,7 @@ function LZ4F_frameInfo_t(;
     blockmode::BlockMode=block_linked,
     contentchecksum::Bool=false,
     frametype::FrameType=normal_frame,
-    contentsize::Int64=0, 
+    contentsize::Int64=Int64(0), 
     blockchecksum::Bool=false,
 )
     LZ4F_frameInfo_t(blocksizeid, blockmode, Cuint(contentchecksum), frametype, contentsize, Cuint(0), Cuint(blockchecksum))
